Rename Rails server to generic
